python/requirements.txt
Cython
numpy
Pillow
scikit-learn
